requirements.txt
Flask==1.0.4
APScheduler==3.7.0
requests==2.26.0
slack-sdk==3.13.0
pymongo
Markupsafe==1.1.1
ItsDangerous==1.1.0
beautifulsoup4==4.10.0
python-dotenv==0.19.0
python-dateutil==2.8.2
pymongo[srv]